requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
beautifulsoup4==4.11.1

channels==4.0.0
channels-redis==4.0.0
confusable-homoglyphs==3.2.0
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==3.2.14
django-bootstrap4==22.2
django-import-export==2.8.0
django-ranged-response==0.2.0
django-recaptcha==3.0.0
django-registration==3.3
django-simple-captcha==0.5.17
et-xmlfile==1.1.0
gunicorn==20.1.0
import-export==0.2.67.dev6
MarkupPy==1.14
msgpack==1.0.4
mysqlclient==2.1.1
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
Pillow==9.2.0
psycopg2==2.9.4
pyparsing==3.0.9
pytz==2022.2.1
PyYAML==6.0

redis==4.3.4
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tablib==3.2.1
whitenoise==6.2.0
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
